vidvibe_test/vidvibe_test/requirements.txt
Flask
openai
yt-dlp
whisper
python-dotenv